modify 32 and 64 bits probing table to use the same layout as used in 128 bits tables

+++ hashprobe/hashprobe_amd64.go
@@ -6,7 +6,7 @@ package hashprobe
 func multiProbe32(table []uint32, len, cap int, hashes []uintptr, keys []uint32, values []int32) int
 
 //go:noescape
-func multiProbe64(table []uint64, len, cap int, hashes []uintptr, keys []uint64, values []int32) int
+func multiProbe64(table []byte, len, cap int, hashes []uintptr, keys []uint64, values []int32) int
 
 //go:noescape
 func multiProbe128(table []byte, len, cap int, hashes []uintptr, keys [][16]byte, values []int32) int
